# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "工程认证智能体 API",
    "summary": "实验教学证据治理与持续改进平台",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/academic/catalog": {
      "get": {
        "tags": [
          "academic"
        ],
        "summary": "Get academic master data catalog",
        "operationId": "get_catalog_api_v1_academic_catalog_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicCatalogResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/academic/program": {
      "patch": {
        "tags": [
          "academic"
        ],
        "summary": "Update current academic program",
        "operationId": "update_program_api_v1_academic_program_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcademicProgramUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicProgramResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/academic/courses": {
      "post": {
        "tags": [
          "academic"
        ],
        "summary": "Create one course",
        "operationId": "create_course_api_v1_academic_courses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/academic/courses/{course_id}": {
      "patch": {
        "tags": [
          "academic"
        ],
        "summary": "Update one course",
        "operationId": "update_course_api_v1_academic_courses__course_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/academic/graduation-requirements": {
      "post": {
        "tags": [
          "academic"
        ],
        "summary": "Create one graduation requirement",
        "operationId": "create_requirement_api_v1_academic_graduation_requirements_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraduationRequirementUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraduationRequirementResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/academic/graduation-requirements/{requirement_id}": {
      "patch": {
        "tags": [
          "academic"
        ],
        "summary": "Update one graduation requirement",
        "operationId": "update_requirement_api_v1_academic_graduation_requirements__requirement_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "requirement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Requirement Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraduationRequirementUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraduationRequirementResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/academic/indicators": {
      "post": {
        "tags": [
          "academic"
        ],
        "summary": "Create one competency indicator",
        "operationId": "create_indicator_api_v1_academic_indicators_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompetencyIndicatorUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyIndicatorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/academic/indicators/{indicator_id}": {
      "patch": {
        "tags": [
          "academic"
        ],
        "summary": "Update one competency indicator",
        "operationId": "update_indicator_api_v1_academic_indicators__indicator_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "indicator_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Indicator Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompetencyIndicatorUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyIndicatorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/academic/objectives": {
      "post": {
        "tags": [
          "academic"
        ],
        "summary": "Create one course objective",
        "operationId": "create_objective_api_v1_academic_objectives_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseObjectiveUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseObjectiveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/academic/objectives/{objective_id}": {
      "patch": {
        "tags": [
          "academic"
        ],
        "summary": "Update one course objective",
        "operationId": "update_objective_api_v1_academic_objectives__objective_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "objective_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Objective Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseObjectiveUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseObjectiveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/academic/experiments": {
      "post": {
        "tags": [
          "academic"
        ],
        "summary": "Create one experiment project",
        "operationId": "create_experiment_api_v1_academic_experiments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperimentProjectUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/academic/experiments/{experiment_id}": {
      "patch": {
        "tags": [
          "academic"
        ],
        "summary": "Update one experiment project",
        "operationId": "update_experiment_api_v1_academic_experiments__experiment_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperimentProjectUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/academic/rubric-items": {
      "post": {
        "tags": [
          "academic"
        ],
        "summary": "Create one rubric item",
        "operationId": "create_rubric_item_api_v1_academic_rubric_items_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RubricItemUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RubricItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/academic/rubric-items/{item_id}": {
      "patch": {
        "tags": [
          "academic"
        ],
        "summary": "Update one rubric item",
        "operationId": "update_rubric_item_api_v1_academic_rubric_items__item_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RubricItemUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RubricItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AcademicCatalogResponse": {
        "properties": {
          "program": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AcademicProgramResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "courses": {
            "items": {
              "$ref": "#/components/schemas/CourseResponse"
            },
            "type": "array",
            "title": "Courses"
          },
          "graduationRequirements": {
            "items": {
              "$ref": "#/components/schemas/GraduationRequirementResponse"
            },
            "type": "array",
            "title": "Graduationrequirements"
          },
          "indicators": {
            "items": {
              "$ref": "#/components/schemas/CompetencyIndicatorResponse"
            },
            "type": "array",
            "title": "Indicators"
          },
          "objectives": {
            "items": {
              "$ref": "#/components/schemas/CourseObjectiveResponse"
            },
            "type": "array",
            "title": "Objectives"
          },
          "experiments": {
            "items": {
              "$ref": "#/components/schemas/ExperimentProjectResponse"
            },
            "type": "array",
            "title": "Experiments"
          },
          "rubricItems": {
            "items": {
              "$ref": "#/components/schemas/RubricItemResponse"
            },
            "type": "array",
            "title": "Rubricitems"
          },
          "sourceMaterials": {
            "items": {
              "$ref": "#/components/schemas/SourceMaterialResponse"
            },
            "type": "array",
            "title": "Sourcematerials"
          },
          "supportLinks": {
            "items": {
              "$ref": "#/components/schemas/SupportLinkResponse"
            },
            "type": "array",
            "title": "Supportlinks"
          }
        },
        "type": "object",
        "required": [
          "program",
          "courses",
          "graduationRequirements",
          "indicators",
          "objectives",
          "experiments",
          "rubricItems",
          "sourceMaterials",
          "supportLinks"
        ],
        "title": "AcademicCatalogResponse"
      },
      "AcademicProgramResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "discipline": {
            "type": "string",
            "title": "Discipline"
          },
          "degree": {
            "type": "string",
            "title": "Degree"
          },
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "evaluationCycle": {
            "type": "string",
            "title": "Evaluationcycle"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "discipline",
          "degree",
          "owner",
          "evaluationCycle",
          "status"
        ],
        "title": "AcademicProgramResponse"
      },
      "AcademicProgramUpsertRequest": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "discipline": {
            "type": "string",
            "title": "Discipline"
          },
          "degree": {
            "type": "string",
            "title": "Degree"
          },
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "evaluationCycle": {
            "type": "string",
            "title": "Evaluationcycle"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "discipline",
          "degree",
          "owner",
          "evaluationCycle",
          "status"
        ],
        "title": "AcademicProgramUpsertRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}